+++ src/vint.hpp
@@ -3,8 +3,8 @@
 
 namespace gen {
 
-static int8_t append_rvint32(byte_vec& vec, uint32_t u32) {
-  int8_t len = 0;
+static size_t append_rvint32(byte_vec& vec, uint32_t u32) {
+  size_t len = 0;
   while (u32 > 31) {
     vec.push_back(u32 & 0xFF);
     u32 >>= 8;
@@ -15,15 +15,15 @@ static int8_t append_rvint32(byte_vec& vec, uint32_t u32) {
 }
 static uint32_t read_rvint32(const uint8_t *ptr) {
   uint32_t ret = *ptr >> 3;
-  int len = *ptr-- & 0x07;
+  size_t len = *ptr-- & 0x07;
   while (len--) {
     ret <<= 7;
     ret += *ptr--;
   }
   return ret;
 }
-static int8_t append_fvint32(byte_vec& vec, uint32_t u32) {
-  int8_t len = 0;
+static size_t append_fvint32(byte_vec& vec, uint32_t u32) {
+  size_t len = 0;
   do {
     uint8_t b = u32 & 0x7F;
     u32 >>= 7;
@@ -34,8 +34,8 @@ static int8_t append_fvint32(byte_vec& vec, uint32_t u32) {
   } while (u32 > 0);
   return len;
 }
-static int8_t copy_fvint32(uint8_t *ptr, uint32_t u32) {
-  int8_t len = 0;
+static size_t copy_fvint32(uint8_t *ptr, uint32_t u32) {
+  size_t len = 0;
   do {
     uint8_t b = u32 & 0x7F;
     u32 >>= 7;
@@ -46,7 +46,7 @@ static int8_t copy_fvint32(uint8_t *ptr, uint32_t u32) {
   } while (u32 > 0);
   return len;
 }
-static uint32_t read_fvint32(const uint8_t *ptr, int8_t& len) {
+static uint32_t read_fvint32(const uint8_t *ptr, size_t& len) {
   uint32_t ret = *ptr & 0x7F;
   len = 1;
   while (*ptr++ & 0x80) {
@@ -56,27 +56,27 @@ static uint32_t read_fvint32(const uint8_t *ptr, int8_t& len) {
   }
   return ret;
 }
-static int8_t get_vlen_of_uint32(uint32_t vint) {
+static size_t get_vlen_of_uint32(uint32_t vint) {
   return vint < (1 << 7) ? 1 : (vint < (1 << 14) ? 2 : (vint < (1 << 21) ? 3 :
           (vint < (1 << 28) ? 4 : 5)));
 }
-static int append_vint32(byte_vec& vec, uint32_t vint) {
-  int len = get_vlen_of_uint32(vint);
-  for (int i = len - 1; i > 0; i--)
+static size_t append_vint32(byte_vec& vec, uint32_t vint) {
+  size_t len = get_vlen_of_uint32(vint);
+  for (size_t i = len - 1; i > 0; i--)
     vec.push_back(0x80 + ((vint >> (7 * i)) & 0x7F));
   vec.push_back(vint & 0x7F);
   return len;
 }
-static int append_vint32(byte_vec& vec, uint32_t vint, int len) {
-  for (int i = len - 1; i > 0; i--)
+static size_t append_vint32(byte_vec& vec, uint32_t vint, size_t len) {
+  for (size_t i = len - 1; i > 0; i--)
     vec.push_back(0x80 + ((vint >> (7 * i)) & 0x7F));
   vec.push_back(vint & 0x7F);
   return len;
 }
-static int append_ovint32(byte_vec& vec, uint32_t vint, int offset_bits, uint8_t or_mask) {
-  int mask_bits = 7 - offset_bits;
+static size_t append_ovint32(byte_vec& vec, uint32_t vint, size_t offset_bits, uint8_t or_mask) {
+  size_t mask_bits = 7 - offset_bits;
   uint8_t mask = (1 << mask_bits) - 1;
-  int len = 0;
+  size_t len = 0;
   uint8_t b = or_mask;
   do {
     b |= (vint & mask);
@@ -91,8 +91,8 @@ static int append_ovint32(byte_vec& vec, uint32_t vint, int offset_bits, uint8_t
   } while (vint > 0);
   return len;
 }
-static uint32_t read_ovint32(const uint8_t *ptr, int8_t& len, int offset_bits) {
-  int mask_bits = 7 - offset_bits;
+static uint32_t read_ovint32(const uint8_t *ptr, size_t& len, size_t offset_bits) {
+  size_t mask_bits = 7 - offset_bits;
   uint8_t mask = (1 << mask_bits) - 1;
   uint32_t ret = *ptr & mask;
   len = 1;
@@ -105,16 +105,16 @@ static uint32_t read_ovint32(const uint8_t *ptr, int8_t& len, int offset_bits) {
   }
   return ret;
 }
-    static int8_t get_svint60_len(int64_t vint) {
+    static size_t get_svint60_len(int64_t vint) {
       vint = abs(vint);
       return vint < (1 << 4) ? 1 : (vint < (1 << 12) ? 2 : (vint < (1 << 20) ? 3 :
               (vint < (1 << 28) ? 4 : (vint < (1LL << 36) ? 5 : (vint < (1LL << 44) ? 6 :
               (vint < (1LL << 52) ? 7 : 8))))));
     }
-    static int read_svint60_len(uint8_t *ptr) {
+    static size_t read_svint60_len(uint8_t *ptr) {
       return 1 + ((*ptr >> 4) & 0x07);
     }
-    static void copy_svint60(int64_t input, uint8_t *out, int8_t vlen) {
+    static void copy_svint60(int64_t input, uint8_t *out, size_t vlen) {
       vlen--;
       long lng = abs(input);
       *out++ = ((lng >> (vlen * 8)) & 0x0F) + (vlen << 4) + (input < 0 ? 0x00 : 0x80);
@@ -126,7 +126,7 @@ static uint32_t read_ovint32(const uint8_t *ptr, int8_t& len, int offset_bits) {
       bool is_neg = true;
       if (*ptr & 0x80)
         is_neg = false;
-      int len = (*ptr >> 4) & 0x07;
+      size_t len = (*ptr >> 4) & 0x07;
       while (len--) {
         ret <<= 8;
         ptr++;
@@ -134,15 +134,15 @@ static uint32_t read_ovint32(const uint8_t *ptr, int8_t& len, int offset_bits) {
       }
       return is_neg ? -ret : ret;
     }
-    static int8_t get_svint61_len(uint64_t vint) {
+    static size_t get_svint61_len(uint64_t vint) {
       return vint < (1 << 5) ? 1 : (vint < (1 << 13) ? 2 : (vint < (1 << 21) ? 3 :
               (vint < (1 << 29) ? 4 : (vint < (1LL << 37) ? 5 : (vint < (1LL << 45) ? 6 :
               (vint < (1LL << 53) ? 7 : 8))))));
     }
-    static int read_svint61_len(uint8_t *ptr) {
+    static size_t read_svint61_len(uint8_t *ptr) {
       return 1 + (*ptr >> 5);
     }
-    static void copy_svint61(uint64_t input, uint8_t *out, int8_t vlen) {
+    static void copy_svint61(uint64_t input, uint8_t *out, size_t vlen) {
       vlen--;
       *out++ = ((input >> (vlen * 8)) & 0x1F) + (vlen << 5);
       while (vlen--)
@@ -150,7 +150,7 @@ static uint32_t read_ovint32(const uint8_t *ptr, int8_t& len, int offset_bits) {
     }
     static uint64_t read_svint61(uint8_t *ptr) {
       uint64_t ret = *ptr & 0x1F;
-      int len = (*ptr >> 5);
+      size_t len = (*ptr >> 5);
       while (len--) {
         ret <<= 8;
         ptr++;
@@ -158,13 +158,13 @@ static uint32_t read_ovint32(const uint8_t *ptr, int8_t& len, int offset_bits) {
       }
       return ret;
     }
-    static int8_t get_svint15_len(uint64_t vint) {
+    static size_t get_svint15_len(uint64_t vint) {
       return vint < (1 << 7) ? 1 : 2;
     }
-    static int read_svint15_len(uint8_t *ptr) {
+    static size_t read_svint15_len(uint8_t *ptr) {
       return 1 + (*ptr >> 7);
     }
-    static void copy_svint15(uint64_t input, uint8_t *out, int8_t vlen) {
+    static void copy_svint15(uint64_t input, uint8_t *out, size_t vlen) {
       vlen--;
       *out++ = ((input >> (vlen * 8)) & 0x7F) + (vlen << 7);
       while (vlen--)
@@ -172,7 +172,7 @@ static uint32_t read_ovint32(const uint8_t *ptr, int8_t& len, int offset_bits) {
     }
     static uint64_t read_svint15(uint8_t *ptr) {
       uint64_t ret = *ptr & 0x7F;
-      int len = (*ptr >> 7);
+      size_t len = (*ptr >> 7);
       while (len--) {
         ret <<= 8;
         ptr++;
